Use 2D Iterator and work on logic in GeoData#canSeeTarget methods.
The 3D version has a flaw that it always picks the highest Z from
repeating (x,y) pairs. This results in wrong Z values being checked in
the loop. This also means that we have the floating point calculation
back in the loop.

Shooting over fences are now only possible if the oponents are far
enough away from each other. This fixes the issue with shooting through
the floor or the roof.

Now you can also shoot over fences near the character on lower ground.