Add YES and NO to the MsgBoxResult enum in the jfx utils.
Add YES and NO to the MsgBoxResult enum to be able to differentiate
between OK and YES and between CANCEL and NO. The reasons are the "(Do
not) Save" buttons in the MsgBoxButton enum. On a MsbBox with "Save",
"Do not Save" and "Cancel" buttons to save unsaved work for example, you
need to differentiate between CANCEL and NO.

However, a better solution must be found so in the future the buttons on
a message box can have a custom, internationalizable text.